group_chat/requirements.txt
a2a-sdk==0.2.11
uvicorn==0.35.0
httpx==0.28.1
python-dotenv
rich
azure-ai-agents
azure-identity
semantic-kernel==1.34.0